+++ include/rte_flow/dp_rte_flow_init.h
@@ -10,10 +10,10 @@ extern "C" {
 #include <rte_flow.h>
 #include "dp_port.h"
 
-int dp_install_isolated_mode_ipip(int port_id, uint8_t proto_id);
+int dp_install_isolated_mode_ipip(uint16_t port_id, uint8_t proto_id);
 
 #ifdef ENABLE_VIRTSVC
-int dp_install_isolated_mode_virtsvc(int port_id, uint8_t proto_id, const uint8_t svc_ipv6[16], uint16_t svc_port);
+int dp_install_isolated_mode_virtsvc(uint16_t port_id, uint8_t proto_id, const uint8_t svc_ipv6[16], uint16_t svc_port);
 #endif
 
 #ifdef __cplusplus
