+++ src/dp_util.c
@@ -12,7 +12,7 @@
 #define DP_SYSFS_MAX_PATH 256
 
 // makes sure there is enough space to prevent collisions
-#define DP_JHASH_MARGIN_COEF(ENTRIES) ((ENTRIES)*1.20)
+#define DP_JHASH_MARGIN_COEF(ENTRIES) ((uint32_t)((ENTRIES)*1.20))
 
 int dp_get_dev_info(uint16_t port_id, struct rte_eth_dev_info *dev_info, char ifname[IF_NAMESIZE])
 {
@@ -138,7 +138,7 @@ struct rte_hash *dp_create_jhash_table(int entries, size_t key_len, const char *
 	struct rte_hash_parameters params = {
 		.name = full_name,
 		.entries = DP_JHASH_MARGIN_COEF(entries),
-		.key_len = key_len,
+		.key_len = (uint32_t)key_len,  // no way this will get bigger than 32b
 		.hash_func = hash_func,
 		.hash_func_init_val = 0xfee1900d,  // "random" IV
 		.socket_id = socket_id,
